# patch: optimize encoding and decoding functions