Currently, the encoding and decoding functions concat strings to then test these strings in switch statements.
After optimizing these function by working on the string directly, `encode` runs 10-20% faster while `decode` runs 30-40% faster.

Timings taken using an array containing 10000x 90kb of json. (`Array(10000).fill(<json>)`):
```
17/09/2025, 15:53:57: load test data: 2 milliseconds
17/09/2025, 15:53:57: prepare json string: 1.21 seconds
17/09/2025, 15:53:58: --------: instantly
17/09/2025, 15:53:58: compress-json[compress]: 2.76 seconds
17/09/2025, 15:54:01: save data/compress-json.json: 1 millisecond
17/09/2025, 15:54:01: compress-json[decompress]: 5.6 seconds
17/09/2025, 15:54:07: compress-json[compare]: 1.35 seconds
17/09/2025, 15:54:08: --------: instantly
17/09/2025, 15:54:08: optimized[compress]: 2.26 seconds
17/09/2025, 15:54:10: save data/optimized.json: instantly
17/09/2025, 15:54:10: optimized[decompress]: 3.5 seconds
17/09/2025, 15:54:14: optimized[compare]: 1.34 seconds
done.
```
Timings taken using an array containing 1000x 4.9mb of json. (`Array(1000).fill(<json>)`):
```
17/09/2025, 16:03:35: load test data: 23 milliseconds
17/09/2025, 16:03:35: prepare json string: 658 milliseconds
17/09/2025, 16:03:35: --------: instantly
17/09/2025, 16:03:35: compress-json[compress]: 1.66 seconds
17/09/2025, 16:03:37: save data/compress-json.json: 3 milliseconds
17/09/2025, 16:03:37: compress-json[decompress]: 3.27 seconds
17/09/2025, 16:03:40: compress-json[compare]: 726 milliseconds
17/09/2025, 16:03:41: --------: instantly
17/09/2025, 16:03:41: optimized[compress]: 1.43 seconds
17/09/2025, 16:03:42: save data/optimized.json: 4 milliseconds
17/09/2025, 16:03:42: optimized[decompress]: 2.28 seconds
17/09/2025, 16:03:45: optimized[compare]: 691 milliseconds
done.
```

_Tests modified to run both old and new code side-by-side_