@beenotung 2025-09-21
The patch is released to npm as v3.4.0 :rocket: