@doxick 2025-09-21
I just saw the changes you made to inline and optimize. Looks great, and great to see you actively working on it as well!

Curious to how much of an impact it has on the benchmarks, before vs after.