@beenotung 2025-09-25
@doxick On a "specially designed test set", the new changes showed about 1.49× faster performance.
However, in real-world usage, the effect will depend heavily on the "data shape", making it harder to predict the actual impact in practice.
That said, the changes generally make the program perform fewer operations, so there should be no downside for performance—just a small, acceptable trade-off in readability.